requirements.txt
# --- FastAPI Core ---
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.8.2

# --- Database ---
SQLAlchemy==2.0.25
psycopg2-binary==2.9.9

# --- Auth & Security ---
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
python-jose[cryptography]==3.3.0
PyJWT==2.12.1
email-validator==2.3.0

# --- File Handling ---
python-docx==1.1.0
pdfplumber==0.11.9
python-multipart==0.0.6
aiofiles==23.2.1

# --- Google APIs ---
google-api-python-client==2.116.0
google-auth-oauthlib==1.2.0
google-auth==2.27.0

# --- Utilities ---
requests==2.33.1
python-dotenv==1.2.2

# --- OpenAI ---
openai==2.31.0

# --- AI / Vector Search (FIXED) ---
numpy==1.26.4
faiss-cpu==1.10.0
sentence-transformers==2.7.0

# --- Compatibility Fix ---
protobuf==4.25.3
nest-asyncio==1.6.0
pgvector==0.3.6